Merge in TROJ/trojai from optimizer_update to master

* commit 'eca578e52545cc0212bab2d129dcf27ddb0b5784':
  fixed unittests, ported over functionality from private to public, and updated package requirements
  fixed unit-tests, updated setup.py to include necessary packages for testing, and updated default soft-to-hard function to be clearer code
  requirements update
  fixed constructors for config
  removed commented out code
  typo
  -fixed code review doc string -moved polygon trigger into datagen as an ImageEntity
  fixed code review issues
  -refactor of adv opimizers -reduced logging level for transformations to debug
  added missing dependency to advertorch
  -added transformations used in round1-3 to public repo -added disable_progress_bar to reporting_config -added support for FBF and PGD

+++ trojai/test/modelgen/test_default_optimizer.py
@@ -114,7 +114,9 @@ class TestRunner(unittest.TestCase):
                             step = 0.1
                             batch_acc_vec = np.arange(0, 1 + step, step)
                             for batch_acc in batch_acc_vec:
-                                random_mat = self.rso.rand(batch_size, num_outputs)
+                                # cast the generated data as float-32 so that there is no loss of accuracy
+                                # between converting from numpy to PyTorch
+                                random_mat = self.rso.rand(batch_size, num_outputs).astype(np.float32)
                                 row_sum = random_mat.sum(axis=1)
 
                                 # normalize the random_mat such that every row adds up to 1
